requirements.txt
pytesseract
pypresence
pillow
pywin32
python-dotenv
tk  # Optional, untuk GUI di beberapa OS non-Windows
